+++ scripts/vast_start_phases.py
@@ -1,3 +1,0 @@
-"""Deprecated: use ``rfdetr_demo.vast.start_phases``."""
-
-from rfdetr_demo.vast.start_phases import *  # noqa: F403