+++ scripts/vast_start_progress.py
@@ -1,3 +1,0 @@
-"""Deprecated: use ``rfdetr_demo.vast.start_progress``."""
-
-from rfdetr_demo.vast.start_progress import *  # noqa: F403